Add files via upload
This is the last data wrangling with the PDO from http://research.jisao.washington.edu/pdo/PDO.latest.txt. While I think this is good data, it's climate change adjust or normalized  in the time series, which makes it not the same as the values measured by NOAA. I'm going to use NOAA from here on out.